Add tests for vadda

+++ code/vadda_b.h
@@ -1,3 +1,3 @@
 for (int i = 0; i < 16; i++) {
-  dst.byte[i] = abs(a.byte[i]) + abs(b.byte[i]);
+  dst.byte[i] = abs((s8)a.byte[i]) + abs((s8)b.byte[i]);
 }
