+++ code/vadda_d.h
@@ -1,4 +1,4 @@
 for (int i = 0; i < 2; i++) {
-  dst.dword[i] = abs(a.dword[i]) + abs(b.dword[i]);
+  dst.dword[i] = abs((s64)a.dword[i]) + abs((s64)b.dword[i]);
 }
