+++ code/vadda_h.h
@@ -1,4 +1,4 @@
 for (int i = 0; i < 8; i++) {
-  dst.half[i] = abs(a.half[i]) + abs(b.half[i]);
+  dst.half[i] = abs((s16)a.half[i]) + abs((s16)b.half[i]);
 }
