+++ code/vadda_w.h
@@ -1,3 +1,3 @@
 for (int i = 0; i < 4; i++) {
-  dst.word[i] = abs(a.word[i]) + abs(b.word[i]);
+  dst.word[i] = abs((s32)a.word[i]) + abs((s32)b.word[i]);
 }
